WIP - Fix shader compile errors
Hardcode shader paths - Remove this! This is what I’m going to focus on
now but want a “checkpoint” commit first.

Used simpler “test” shaders just to test.

App doesn’t react to keyboard input, and the main.cpp doesn’t match the
tutorial so it’s stuck in a loop instead of progressing through steps.