Use C++11 raw string literals to include shaders
The shader’s need to be included in the build somehow otherwise we
can’t ship the project as a binary, and it’s won’t even run from CLion
as the actual executable isn’t in the folder where the source files are
so the shader paths were failing.

Solving this with an interesting hack/idea with C++11 raw string
literals, which puts R”(…)” around the shader source code in the shader
files, and then pulling them in at compile time through #include. Idea
gotten from:
http://stackoverflow.com/a/33172917