Add 3D OpenGL sample
Following the same tutorial series as previously, this is part 4.
http://headerphile.com/sdl2/opengl-part-4-3d-basics/

Just minimal changes to make it work on OSX and not using GLEW.

Note: This doesn’t work from CLion as the shader files aren’t where the
executable is running from so they fail to load. I can solve this with
the C++11 raw string literals, but I’ll commit that separately.