Use C++11 raw string literals to include data files
The shaders and model data need to be included in the build somehow otherwise we can’t ship the project as a binary, and it won’t work from CLion as the running executable isn’t in the project root folder so no files are found when loading.

Solving this with an interesting hack/idea with C++11 raw string literals, which puts R”(…)” around the data files, and then pulling them in at compile time through #include. Reference:
http://stackoverflow.com/a/33172917