Add glm library
"OpenGL Mathematics (GLM) is a header only C++ mathematics library for
graphics software based on the OpenGL Shading Language (GLSL)
specification and released under the MIT license."

We need it for vector and matrix operations in our camera, shaders, and
anything that we want to transform and render.

Note: Original commit by Jocke. We just broke down the commit on JP's machine.