Rename project to TurboLoader
"A fast loader is a software program for a home computer, such as the
Commodore 64 or ZX Spectrum, that accelerates the speed of file loading
from floppy disk or compact cassette."
https://en.wikipedia.org/wiki/Fast_loader

TurboTape and TurboDisk were two popular C64 fast loaders. Turbo Pascal
was the first programming language we learned. Street Fighter 2 Turbo
was so blazingly fast that it reinvented the game. And who can forget
the Turbo button on the first PC's?

We want to create something that is blazingly fast for today's best home
computer, the Raspberry Pi. So "TurboLoader" is a fitting name for an
emulator and ROM launcher of all the fun games we grew up with.